Updated spell check target and refactored package name
