+++ src/main/java/com/github/onsdigital/babbage/search/external/requests/search/requests/ContentQuery.java
@@ -3,8 +3,8 @@ package com.github.onsdigital.babbage.search.external.requests.search.requests;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.github.onsdigital.babbage.configuration.Configuration;
 import com.github.onsdigital.babbage.search.external.SearchType;
-import com.github.onsdigital.babbage.search.external.requests.suggest.SpellCheckRequest;
-import com.github.onsdigital.babbage.search.external.requests.suggest.models.SpellingCorrection;
+import com.github.onsdigital.babbage.search.external.requests.spellcheck.SpellCheckRequest;
+import com.github.onsdigital.babbage.search.external.requests.spellcheck.models.SpellingCorrection;
 import com.github.onsdigital.babbage.search.input.SortBy;
 import com.github.onsdigital.babbage.search.input.TypeFilter;
 import com.github.onsdigital.babbage.search.model.ContentType;
